refactor(db): abstract initLISHsTables into own module
